Accept decimals for paralog_tolerance

+++ captus/captus_assembly.py
@@ -663,8 +663,8 @@ class CaptusAssembly(object):
         output_group.add_argument(
             "--paralog_tolerance",
             action="store",
-            default=5,
-            type=int,
+            default=5.0,
+            type=float,
             dest="paralog_tolerance",
             help="Only paralogs with a wscore >= (locus best hit wscore / paralog_tolerance) will be"
             " retained",
